+++ lib/env.ex
@@ -4,6 +4,6 @@ defmodule EctoCooler.Env do
 
   defp safe_get(:error, default), do: default
   defp safe_get({:ok, value}, _), do: value
-  defp safe_get({:error, _}), do: nil
+  defp safe_get(:error), do: nil
   defp safe_get({:ok, value}), do: value
 end