+++ lib/option_parser.ex
@@ -5,7 +5,7 @@ defmodule EctoCooler.OptionParser do
 
   @functions [
     {"all", 1},
-    {"change", 1},
+    {"change", 2},
     {"changeset", 0},
     {"create!", 1},
     {"create", 1},
@@ -21,17 +21,14 @@ defmodule EctoCooler.OptionParser do
 
   @spec parse(String.t(), list() | atom()) :: map()
 
-  def parse(suffix, []) do
-    @functions
-    |> Enum.reduce(%{}, fn {function, arity}, acc ->
-      Map.put(acc, String.to_atom(function), %{
-        name: function_name(function, suffix),
-        description: function_description(function, arity, suffix)
-      })
-    end)
-  end
-
   def parse(suffix, :read), do: parse(suffix, only: [:all, :get, :get!, :get_by, :get_by!])
+  def parse(suffix, :delete), do: parse(suffix, only: [:delete, :delete!])
+
+  def parse(suffix, :write) do
+    parse(suffix,
+      only: [:change, :changeset, :create, :create!, :update, :update!, :delete, :delete!]
+    )
+  end
 
   def parse(suffix, :read_write) do
     parse(suffix,
@@ -65,72 +62,30 @@ defmodule EctoCooler.OptionParser do
   @spec create_suffix(module, list()) :: String.t()
   def create_suffix(_, options) when not is_list(options), do: ""
 
-  def create_suffix(schema, options) when is_list(options) do
-    if Keyword.get(options, :suffix, false) == false do
-      ""
-    else
-      Helpers.underscore_module_name(schema)
-    end
+  def create_suffix(schema, options) do
+    if Keyword.get(options, :suffix, false),
+      do: Helpers.underscore_module_name(schema),
+      else: ""
   end
 
-  def create_suffix(schema, _), do: Helpers.underscore_module_name(schema)
+  defp build_function_name("all", suffix), do: "all_" <> Drops.Inflector.pluralize(suffix)
+  defp build_function_name("get_by", suffix), do: "get_#{suffix}_by"
+  defp build_function_name("get_by!", suffix), do: "get_#{suffix}_by!"
+  defp build_function_name("changeset", suffix), do: "#{suffix}_changeset"
+
+  defp build_function_name(function, suffix) do
+    if String.ends_with?(function, "!"),
+      do: String.replace_suffix(function, "!", "") <> "_" <> suffix <> "!",
+      else: function <> "_" <> suffix
+  end
 
   defp function_name(function, ""), do: String.to_atom(function)
+  defp function_name(function, suffix), do: String.to_atom(build_function_name(function, suffix))
 
-  defp function_name("all", suffix),
-    do: String.to_atom("all_" <> Drops.Inflector.pluralize(suffix))
+  defp function_description(function, arity, ""), do: function <> "/" <> Integer.to_string(arity)
 
-  defp function_name("get_by", suffix) do
-    String.to_atom("get_#{suffix}_by")
-  end
-
-  defp function_name("get_by!", suffix) do
-    String.to_atom("get_#{suffix}_by!")
-  end
-
-  defp function_name("changeset", suffix) do
-    String.to_atom("#{suffix}_changeset")
-  end
-
-  defp function_name(function, suffix) do
-    case function =~ ~r/!/ do
-      true -> String.replace_suffix(function, "!", "") <> "_" <> suffix <> "!"
-      false -> function <> "_" <> suffix
-    end
-    |> String.to_atom()
-  end
-
-  defp function_description(function, arity, "") do
-    function <> "/" <> Integer.to_string(arity)
-  end
-
-  defp function_description("all", arity, suffix) do
-    "all_" <> Drops.Inflector.pluralize(suffix) <> "/" <> Integer.to_string(arity)
-  end
-
-  defp function_description("get_by", arity, suffix) do
-    arity = Integer.to_string(arity)
-    "get_" <> suffix <> "_by" <> "/" <> arity
-  end
-
-  defp function_description("get_by!", arity, suffix) do
-    arity = Integer.to_string(arity)
-    "get_" <> suffix <> "_by!" <> "/" <> arity
-  end
-
-  defp function_description("changeset", arity, suffix) do
-    arity = Integer.to_string(arity)
-    suffix <> "_changeset" <> "/" <> arity
-  end
-
-  defp function_description(function, arity, suffix) do
-    arity = Integer.to_string(arity)
-
-    case function =~ ~r/!/ do
-      true -> String.replace_suffix(function, "!", "") <> "_" <> suffix <> "!" <> "/" <> arity
-      false -> function <> "_" <> suffix <> "/" <> arity
-    end
-  end
+  defp function_description(function, arity, suffix),
+    do: build_function_name(function, suffix) <> "/" <> Integer.to_string(arity)
 
   defp filter_functions(functions, options) when is_list(options) do
     filter_functions(functions, Enum.into(options, %{}))