+++ test/ecto_cooler/with_suffix_test.exs
@@ -63,7 +63,7 @@ defmodule EctoCooler.WithoutSuffixTest do
   end
 
   describe "create" do
-    test "with valid attributbes, it creates a new record" do
+    test "with valid attributes, it creates a new record" do
       {:ok, person} = People.create_person(@person_attributes)
 
       assert Repo.all(Person) == [person]
@@ -75,7 +75,7 @@ defmodule EctoCooler.WithoutSuffixTest do
   end
 
   describe "create!" do
-    test "whith valid attributes, it creates a new record" do
+    test "with valid attributes, it creates a new record" do
       person = People.create_person!(@person_attributes)
 
       assert Repo.all(Person) == [person]